apps/user/urls.py
from rest_framework import routers
from . import views
from django.conf.urls import url, include
from django.urls import path

# from cmt.users.views import (
#     user_list_view,
#     user_redirect_view,
#     user_update_view,
#     user_detail_view,
# )

app_name = "user"

router = routers.DefaultRouter()
router.register(r'user', views.UserViewSet)

urlpatterns = [
    # path("", view=user_list_view, name="list"),
    # path("~redirect/", view=user_redirect_view, name="redirect"),
    # path("~update/", view=user_update_view, name="update"),
    # path("<str:username>/", view=user_detail_view, name="detail"),
    url(r'api/', include(router.urls)),
]
