requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.4
django-allauth==0.44.0
django-codemirror2==0.2
django-cors-headers==3.7.0
django-cte==1.1.4
django-environ==0.4.5
django-import-export==2.4.0
django-model-utils==4.1.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
idna==2.10
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
nameparser==1.0.6
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
pip-autoremove==0.9.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
